backend/requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
Django==3.2.6
django-ckeditor==6.1.0
django-cors-headers==3.8.0
django-debug-toolbar==3.2.2
django-js-asset==1.2.2
django-json-field-schema-validator==0.0.2
django-prettyjson==0.4.1
djangorestframework==3.12.4
flake8==3.9.2
flake8-broken-line==0.3.0
flake8-isort==4.0.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
flower==1.0.0
gunicorn==20.1.0
h2==2.6.2
hpack==3.0.0
humanize==3.11.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.2
isort==5.9.3
jsonschema==3.2.0
kombu==5.1.0
mccabe==0.6.1
mysqlclient==2.0.3
pep8-naming==0.12.1
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
standardjson==0.3.1
testfixtures==6.18.1
tornado==6.1
ua-parser==0.10.0
urllib3==1.26.6
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
